CUPS architecture for VPP and StrongSwan
Separate out the StrongSwan control plane from the VPP dataplane. Note, these
are still tied together due to the VPP API. The VPP API uses shared memory,
so this shared memory needs to be shared across the containers. Also, the
punt sockets for IKE packets from the dataplane to StrongSwan need to be
handled. socat is used for this for now.

Signed-off-by: Kyle Mestery <mestery@mestery.com>