CUPS Implementation of StrongSwan+VPP IPsec VPN (#28)
* CUPS architecture for VPP and StrongSwan

Separate out the StrongSwan control plane from the VPP dataplane. Note, these
are still tied together due to the VPP API. The VPP API uses shared memory,
so this shared memory needs to be shared across the containers. Also, the
punt sockets for IKE packets from the dataplane to StrongSwan need to be
handled. socat is used for this for now.

Signed-off-by: Kyle Mestery <mestery@mestery.com>

* Hook cups-vppvpn into top-level Makefile

Signed-off-by: Kyle Mestery <mestery@mestery.com>

* Change IP addresses for CUPS setup

These share the same network as the regular vppvpn containers, so they
cannot have the same IP addresses.

Signed-off-by: Kyle Mestery <mestery@mestery.com>

* Sleep for a bit

Signed-off-by: Kyle Mestery <mestery@mestery.com>

* Debug ping failure in CI

Signed-off-by: Kyle Mestery <mestery@mestery.com>

* Allow the CUPS and vppvpn tests to run on the same Docker host

Use different Docker networks, as well as different IP subnets.

Signed-off-by: Kyle Mestery <mestery@mestery.com>